Fix for dict unfolding
Dict unfolding (_search_dict function) now operates with array of items
to search. Order of elements in list - is the order of unfolding the
dicts.
For example to unfold {'a':{'b':'c'}}, we must pass an array ['a','b']
Passing full path as aruments required for the following situation.
test ={'root':
         {'x':
              {'test': True},
          'y':
              {'test': False}
          }
     }
Before this path  we passed only 'test' to the search_dict functino and first math has
been returned. This was wrong for example for get('x.y.test') because
True was returned.

Test added.